% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounded_return.R
\name{compounded_return}
\alias{compounded_return}
\title{Calculate compounded returns}
\usage{
compounded_return(price_vector)
}
\arguments{
\item{price_vector}{A vector of numeric prices.}
}
\value{
A vector of compounded returns.
}
\description{
This function takes a vector of prices to calculate compounded return for each time unit.
}
\examples{
prices <- c(1, 1.1, 1.3, 1.2, 1.5, 1.9, 1.8, 2.0)
compounded_return(prices)
}
