% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indice_list}
\alias{indice_list}
\title{8 indices from 1999 to 2016}
\format{A list of 8 data frames with 4550 observations of 3 variables respectively:
\describe{
  \item{`Index name`}{Date, from 1999-01-04 to 2016-12-08}
  \item{Price}{price, in US dollars}
  \item{Comp_Return}{Compounded return, ratio}
}}
\source{
Class material of Columbia IEOR Big Data in Finance
}
\usage{
indice_list
}
\description{
A dataset containing the daily prices and compounded returns of 8 indices from 1999-01-04 to 2016-12-08.
}
\keyword{datasets}
