% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounded_return_visualization.R
\name{return_viz}
\alias{return_viz}
\title{Visualize the compounded returns}
\usage{
return_viz(return_vector, period = c(1:length(return_vector)))
}
\arguments{
\item{return_vector}{A vector of compounded returns.}

\item{period}{A vector of time period. Default as 1:length(return_vector)}
}
\value{
A line plot of compounded returns over the period.
}
\description{
This function uses `ggplot2` package to demonstrate a line plot of compounded returns in the period.
}
\details{
Use `theme_classic()` for styling.
}
\examples{
prices <- c(1, 1.1, 1.3, 1.2, 1.5, 1.9, 1.8, 2.0)
returns <- compounded_return(prices)
return_viz(returns)
}
